Remove controls mappings implementation
Control mapping is out of the scope of this library. So i refactored
the mouse and keyboard callbacks to simply dispatch events when an event
handler as been configured for the key stroke combination.

In order to register an event handler for user controls, we use a vector
instead of a keyword.